+++ core/src/main/java/hudson/model/Queue.java
@@ -120,7 +120,7 @@ import org.kohsuke.accmod.restrictions.DoNotUse;
 import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.converters.basic.AbstractSingleValueConverter;
 import javax.annotation.CheckForNull;
-import jenkins.model.queue.Executable2;
+import jenkins.model.queue.AsynchronousExecution;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.interceptor.RequirePOST;
 
@@ -1479,9 +1479,10 @@ public class Queue extends ResourceController implements Saveable {
         @Nonnull SubTask getParent();
 
         /**
-         * Called by {@link Executor} to perform the task
+         * Called by {@link Executor} to perform the task.
+         * @throws AsynchronousExecution if you would like to continue without consuming a thread
          */
-        @Override void run();
+        @Override void run() throws AsynchronousExecution;
 
         /**
          * Estimate of how long will it take to execute this executable.
